@@ -12,52 +12,52 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "8.2.12",
-    "@angular/common": "8.2.12",
-    "@angular/compiler": "8.2.12",
-    "@angular/core": "8.2.12",
-    "@angular/forms": "8.2.12",
-    "@angular/platform-browser": "8.2.12",
-    "@angular/platform-browser-dynamic": "8.2.12",
-    "@angular/platform-server": "8.2.12",
-    "@angular/router": "8.2.12",
+    "@angular/animations": "11.2.5",
+    "@angular/common": "11.2.5",
+    "@angular/core": "11.2.5",
+    "@angular/forms": "11.2.5",
+    "@angular/platform-browser": "11.2.5",
+    "@angular/platform-browser-dynamic": "11.2.5",
+    "@angular/platform-server": "11.2.5",
+    "@angular/router": "11.2.5",
     "@nguniversal/module-map-ngfactory-loader": "8.1.1",
-    "aspnet-prerendering": "^3.0.1",
-    "bootstrap": "^4.6.0",
-    "chart.js": "^2.9.4",
-    "core-js": "^3.9.1",
-    "jquery": "^3.5.1",
-    "ng2-charts": "^2.4.2",
-    "node-sass": "^5.0.0",
-    "oidc-client": "^1.11.3",
-    "popper.js": "^1.16.0",
-    "rxjs": "^6.6.3",
-    "ts-node": "~8.4.1",
-    "tslint": "~5.20.0",
-    "zone.js": "0.9.1"
+    "aspnet-prerendering": "3.0.1",
+    "bootstrap": "4.6.0",
+    "chart.js": "2.9.4",
+    "core-js": "3.9.1",
+    "jquery": "3.5.1",
+    "ng2-charts": "2.4.2",
+    "node-sass": "5.0.0",
+    "oidc-client": "1.11.3",
+    "popper.js": "1.16.0",
+    "rxjs": "6.6.3",
+    "tslint": "5.20.0",
+    "ts-node": "8.4.1",
+    "zone.js": "0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.803.26",
-    "@angular/cli": "^11.2.4",
-    "@angular/compiler-cli": "^11.2.5",
-    "@angular/language-service": "^8.2.14",
-    "@types/jasmine": "~3.4.4",
-    "@types/jasminewd2": "~2.0.8",
-    "@types/node": "~12.11.6",
-    "codelyzer": "^5.2.2",
-    "jasmine-core": "~3.5.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^5.2.3",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~2.1.0",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.5.4",
-    "typescript": "3.5.3"
+    "@angular-devkit/build-angular": "^0.1102.4",
+    "@angular/cli": "11.2.4",
+    "@angular/compiler": "^11.2.5",
+    "@angular/compiler-cli": "11.2.5",
+    "@angular/language-service": "11.2.5",
+    "@types/jasmine": "3.4.4",
+    "@types/jasminewd2": "2.0.8",
+    "@types/node": "12.11.6",
+    "codelyzer": "5.2.2",
+    "jasmine-core": "3.5.0",
+    "jasmine-spec-reporter": "4.2.1",
+    "karma": "5.2.3",
+    "karma-chrome-launcher": "3.1.0",
+    "karma-coverage-istanbul-reporter": "2.1.0",
+    "karma-jasmine": "2.0.1",
+    "karma-jasmine-html-reporter": "1.5.4",
+    "typescript": "4.1.5"
   },
   "optionalDependencies": {
-    "node-sass": "^5.0.0",
-    "protractor": "^7.0.0",
-    "ts-node": "~8.4.1",
-    "tslint": "~5.20.0"
+    "node-sass": "5.0.0",
+    "protractor": "7.0.0",
+    "ts-node": "8.4.1",
+    "tslint": "5.20.0"
   }
 }
