@@ -1,8 +1,10 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
-import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
+import { HttpClientModule } from '@angular/common/http';
 import { RouterModule } from '@angular/router';
+
+import { ChartsModule } from 'ng2-charts';
 
 import { AppComponent } from './app.component';
 import { NavbarComponent } from './navbar/navbar.component';
@@ -10,6 +12,9 @@ import { SidebarComponent } from './sidebar/sidebar.component';
 import { SectionOrdersComponent } from './sections/section-orders/section-orders.component';
 import { SectionSalesComponent } from './sections/section-sales/section-sales.component';
 import { SectionHealthComponent } from './sections/section-health/section-health.component';
+import { BarChartComponent } from './charts/bar-chart/bar-chart.component';
+import { LineChartComponent } from './charts/line-chart/line-chart.component';
+import { PieChartComponent } from './charts/pie-chart/pie-chart.component';
 
 @NgModule({
   declarations: [
@@ -18,19 +23,22 @@ import { SectionHealthComponent } from './sections/section-health/section-health
     SidebarComponent,
     SectionOrdersComponent,
     SectionSalesComponent,
-    SectionHealthComponent
+    SectionHealthComponent,
+    BarChartComponent,
+    LineChartComponent,
+    PieChartComponent
   ],
   imports: [
     BrowserModule.withServerTransition({ appId: 'ng-cli-universal' }),
     HttpClientModule,
     FormsModule,
+    ChartsModule,
     RouterModule.forRoot([
-      { path: '', redirectTo: '/sales', pathMatch: 'full' },
-
-      { path: 'sales', component: SectionSalesComponent},
-      { path: 'orders', component: SectionOrdersComponent},
-      { path: 'health', component: SectionHealthComponent}
-    ])
+    { path: '', redirectTo: '/sales', pathMatch: 'full' },
+    { path: 'sales', component: SectionSalesComponent },
+    { path: 'orders', component: SectionOrdersComponent },
+    { path: 'health', component: SectionHealthComponent }
+], { relativeLinkResolution: 'legacy' })
   ],
   providers: [],
   bootstrap: [AppComponent]
