@@ -12,12 +12,16 @@ var core_1 = require("@angular/core");
 var forms_1 = require("@angular/forms");
 var http_1 = require("@angular/common/http");
 var router_1 = require("@angular/router");
+var ng2_charts_1 = require("ng2-charts");
 var app_component_1 = require("./app.component");
 var navbar_component_1 = require("./navbar/navbar.component");
 var sidebar_component_1 = require("./sidebar/sidebar.component");
 var section_orders_component_1 = require("./sections/section-orders/section-orders.component");
 var section_sales_component_1 = require("./sections/section-sales/section-sales.component");
 var section_health_component_1 = require("./sections/section-health/section-health.component");
+var bar_chart_component_1 = require("./charts/bar-chart/bar-chart.component");
+var line_chart_component_1 = require("./charts/line-chart/line-chart.component");
+var pie_chart_component_1 = require("./charts/pie-chart/pie-chart.component");
 var AppModule = /** @class */ (function () {
     function AppModule() {
     }
@@ -29,18 +33,22 @@ var AppModule = /** @class */ (function () {
                 sidebar_component_1.SidebarComponent,
                 section_orders_component_1.SectionOrdersComponent,
                 section_sales_component_1.SectionSalesComponent,
-                section_health_component_1.SectionHealthComponent
+                section_health_component_1.SectionHealthComponent,
+                bar_chart_component_1.BarChartComponent,
+                line_chart_component_1.LineChartComponent,
+                pie_chart_component_1.PieChartComponent
             ],
             imports: [
                 platform_browser_1.BrowserModule.withServerTransition({ appId: 'ng-cli-universal' }),
                 http_1.HttpClientModule,
                 forms_1.FormsModule,
+                ng2_charts_1.ChartsModule,
                 router_1.RouterModule.forRoot([
                     { path: '', redirectTo: '/sales', pathMatch: 'full' },
                     { path: 'sales', component: section_sales_component_1.SectionSalesComponent },
                     { path: 'orders', component: section_orders_component_1.SectionOrdersComponent },
                     { path: 'health', component: section_health_component_1.SectionHealthComponent }
-                ])
+                ], { relativeLinkResolution: 'legacy' })
             ],
             providers: [],
             bootstrap: [app_component_1.AppComponent]
