@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { SectionHealthComponent } from './section-health.component';
 
@@ -6,7 +6,7 @@ describe('SectionHealthComponent', () => {
   let component: SectionHealthComponent;
   let fixture: ComponentFixture<SectionHealthComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ SectionHealthComponent ]
     })
