@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { SectionSalesComponent } from './section-sales.component';
 
@@ -6,7 +6,7 @@ describe('SectionSalesComponent', () => {
   let component: SectionSalesComponent;
   let fixture: ComponentFixture<SectionSalesComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ SectionSalesComponent ]
     })
